requirements.txt
# Core AI/LLM Framework
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-google-genai
llama-index-llms-google-genai
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-anthropic
openai
anthropic
google-genai

# UI & PDF Handling 
gradio
PyMuPDF # fitz
pypdf

# Data Exploration & Logging 
datasette

# Vector Store
# chromadb

# Environment & Notebooks
python-dotenv
jupyter # or ipython

# Deployment (Optional - for Modal)
# modal

# PyData essentials
seaborn
matplotlib
numpy
pandas
